+++ src/components/EventOverview/EventOverview.js
@@ -1,59 +1,20 @@
-import axios from 'axios';
-import moment from 'moment';
 import React, { Component } from 'react';
 import { ScrollView, Text, View } from 'react-native';
+
+import axios from 'axios';
+import moment from 'moment';
 import Toast from 'react-native-easy-toast';
 import { connect } from 'react-redux';
-import URL from '../../config';
-import * as actionTypes from '../../store/actions';
+
 import EventImageHeader from '../EventImageHeader/EventImageHeader';
 import Footer from '../Footer/Footer';
-import toasterStyle from '../GeneralStyle/ToasterStyle.style.js';
 import Header from '../Header/Header';
 import HeadlineOverview from '../HeadlineOverview/HeadlineOverview';
+
+import URL from '../../config';
+import * as actionTypes from '../../store/actions';
+import toasterStyle from '../GeneralStyle/ToasterStyle.style.js';
 import styles from './EventOverview.style.js';
-import store from '../../store/store';
-
-import NavigationService from '../../navigation/NavigationService';
-
-axios.interceptors.request.use(request => {
-    const state = store.getState();
-    console.log('request', request);
-    request.headers.Authorization = "Bearer " + state.accessToken;
-    return request;
-})
-axios.interceptors.response.use(response => {
-    return response;
-}, error => {
-    const originalRequest = error.config;
-    if (error.response.status === 401 && !originalRequest.headers.retry) {
-        const state = store.getState();
-        originalRequest.headers.retry = true;
-
-        return axios.post(URL + 'tokens/refresh', {
-            refreshToken: state.refreshToken,
-            userId: state.userID,
-        })
-            .then((response) => {
-                store.dispatch({
-                    type: actionTypes.SAVE_TOKENS, payload: {
-                        accessToken: response.data.accessToken,
-                        refreshToken: response.data.refreshToken
-                    }
-                })
-                return axios(originalRequest);
-            })
-            .catch((error) => {
-                NavigationService.navigate('LoginRoute', {
-                    showErrorMessage: true
-                });
-                return Promise.reject(error);
-            })
-    }
-    return Promise.reject(error);
-})
-
-
 
 class EventOverview extends Component {
